Isometric Test Scene and Pursuit
-Added an isometric test scene in 3d
(Includes a cop and player model)
-Added a pursuit script to be used for the cops.

+++ Assets/Scripts/Map.cs
@@ -6,7 +6,7 @@ public class Map : MonoBehaviour
 {
     [SerializeField] private int Width;
     [SerializeField] private int Height;
-    [SerializeField] private int CellSize;
+    [SerializeField] private float CellSize;
     [SerializeField] private Vector3 Origin;
 
     public Grid<int> MapGrid;
